+++ pyviews/binding/inline.py
@@ -3,10 +3,10 @@
 from functools import partial
 
 from pyviews.binding.binder import BindingContext
-from pyviews.core.bindable import InheritedDict
 from pyviews.core.binding import Binding, BindingCallback, BindingError
 from pyviews.core.error import PyViewsError, error_handling
 from pyviews.core.expression import Expression, execute
+from pyviews.core.rendering import NodeGlobals
 
 
 class InlineBinding(Binding):
@@ -17,7 +17,7 @@ class InlineBinding(Binding):
         callback: BindingCallback,
         bind_expression: Expression,
         value_expression: Expression,
-        expr_vars: InheritedDict
+        expr_vars: NodeGlobals
     ):
         super().__init__()
         self._callback: BindingCallback = callback
@@ -30,13 +30,13 @@ class InlineBinding(Binding):
     def bind(self):
         self.destroy()
         with error_handling(BindingError('Error occurred during inline binding'), self._add_error_info):
-            bind = execute(self._bind_expression, self._expression_vars.to_dictionary())
+            bind = execute(self._bind_expression, self._expression_vars)
             self._destroy = bind(self._execute_callback)
         self._execute_callback()
 
     def _execute_callback(self):
         with error_handling(BindingError, self._add_error_info):
-            value = execute(self._value_expression, self._expression_vars.to_dictionary())
+            value = execute(self._value_expression, self._expression_vars)
             self._callback(value)
 
     def _add_error_info(self, error: PyViewsError):