+++ pyviews/binding/tests/expression_tests.py
@@ -1,13 +1,13 @@
 from unittest.mock import Mock, call
 
-from pytest import fixture, mark, raises
+from pytest import fixture, mark
 
 from pyviews.binding.binder import BindingContext
-from pyviews.binding.expression import ExpressionBinding, bind_setter_to_expression, get_expression_callback
+from pyviews.binding.expression import ExpressionBinding, bind_setter_to_expression
 from pyviews.binding.tests.common import InnerViewModel, ParentViewModel
-from pyviews.core.bindable import InheritedDict
-from pyviews.core.binding import BindingError
+from pyviews.core.bindable import BindableDict
 from pyviews.core.expression import Expression, execute
+from pyviews.core.rendering import NodeGlobals
 from pyviews.core.xml import XmlAttr
 
 
@@ -27,7 +27,7 @@ class ExpressionBindingTests:
 
     callback: Mock
 
-    def _bind(self, expression: Expression, global_vars: InheritedDict) -> ExpressionBinding:
+    def _bind(self, expression: Expression, global_vars: NodeGlobals) -> ExpressionBinding:
         binding = ExpressionBinding(self.callback, expression, global_vars)
         binding.bind()
 
@@ -46,7 +46,7 @@ class ExpressionBindingTests:
         """Target should be updated with expression value on Binding.bind() call"""
         expression = Expression(source)
 
-        self._bind(expression, InheritedDict(global_dict))
+        self._bind(expression, NodeGlobals(global_dict))
 
         expected = execute(expression, global_dict)
         assert self.callback.call_args == call(expected)
@@ -88,12 +88,12 @@ class ExpressionBindingTests:
     def test_expression_changed(self, source: str, global_dict: dict, change):
         """Target should be updated after expression result is changed"""
         expression = Expression(source)
-        global_vars = InheritedDict(global_dict)
+        global_vars = NodeGlobals(global_dict)
         self._bind(expression, global_vars)
 
         change(global_vars)
 
-        expected = execute(expression, global_vars.to_dictionary())
+        expected = execute(expression, global_vars)
         assert self.callback.call_args_list[1:] == [call(expected)]
 
     @mark.parametrize('source, global_dict, change', [
@@ -133,7 +133,7 @@ class ExpressionBindingTests:
     def test_destroy(self, source: str, global_dict: dict, change):
         """Destroy should stop handling expression changes and update target"""
         expression = Expression(source)
-        global_vars = InheritedDict(global_dict)
+        global_vars = NodeGlobals(global_dict)
         binding = self._bind(expression, global_vars)
         self.callback.reset_mock()
 
@@ -151,7 +151,7 @@ def binding_context_fixture(request):
             'setter': setter,
             'xml_attr': xml_attr,
             'expression_body': '1+1',
-            'node': Mock(node_globals = InheritedDict())
+            'node': Mock(node_globals = NodeGlobals())
         }
     )
 
@@ -165,7 +165,7 @@ class BindSetterToExpressionTests:
 
     def test_binds_setter_to_expression_changes(self):
         """should bind setter to expression changes"""
-        self.context.node = Mock(node_globals = InheritedDict({'value': 1}))
+        self.context.node = Mock(node_globals = BindableDict({'value': 1}))
         self.context.expression_body = 'value'
 
         bind_setter_to_expression(self.context)
@@ -179,44 +179,3 @@ class BindSetterToExpressionTests:
         actual = bind_setter_to_expression(self.context)
 
         assert isinstance(actual, ExpressionBinding)
-
-class GetPropertyExpressionCallbackTests:
-    """get_property_expression_callback() tests"""
-
-    @staticmethod
-    @mark.parametrize('expression', ['', '[vm, 2]', 'vm.int_value + val'])
-    def test_raises(expression):
-        """Should raise BindingError if expression is not property expression"""
-        with raises(BindingError):
-            get_expression_callback(Expression(expression), InheritedDict())
-
-    @staticmethod
-    @mark.parametrize('expression_body, value, globals_dict', [
-        ('vm.int_value', 10, {'vm': ParentViewModel(0, InnerViewModel(0, ''))}),
-        ('vm.inner_vm.int_value', 134, {'vm': ParentViewModel(0, InnerViewModel(0, ''))}),
-        ('vm.inner_vm.str_value', 'value', {'vm': ParentViewModel(0, InnerViewModel(0, ''))}),
-        ('vm.inner_vm', InnerViewModel(50, 'a'), {'vm': ParentViewModel(0, InnerViewModel(0, ''))})
-    ]) # yapf: disable
-    def test_callback_updates_property(expression_body, value, globals_dict):
-        """returned callback should update target property"""
-        expression, global_vars = Expression(expression_body), InheritedDict(globals_dict)
-        callback = get_expression_callback(expression, global_vars)
-
-        callback(value)
-
-        assert execute(expression, global_vars.to_dictionary()) == value
-
-    @staticmethod
-    @mark.parametrize('key, value', [
-        ('int_value', 10),
-        ('vm', InnerViewModel(50, 'a')),
-        ('str_value', 'some value')
-    ]) # yapf: disable
-    def test_callback_updates_global_value(key, value):
-        """returned callback should update target property"""
-        global_vars = InheritedDict({key: None})
-        callback = get_expression_callback(Expression(key), global_vars)
-
-        callback(value)
-
-        assert global_vars[key] == value