+++ pyviews/binding/tests/inline_tests.py
@@ -6,9 +6,9 @@ from pytest import fixture, mark
 from pyviews.binding.binder import BindingContext
 from pyviews.binding.inline import InlineBinding, bind_inline
 from pyviews.binding.tests.common import InnerViewModel
-from pyviews.core.bindable import InheritedDict
 from pyviews.core.binding import BindingCallback
 from pyviews.core.expression import Expression
+from pyviews.core.rendering import NodeGlobals
 from pyviews.core.xml import XmlAttr
 
 
@@ -24,7 +24,7 @@ def inline_binding_fixture(request):
 
         return destroy
 
-    expr_globals = InheritedDict({'vm': source_instance, 'bind': bind})
+    expr_globals = NodeGlobals({'vm': source_instance, 'bind': bind})
     bind_expression = Expression('bind')
     value_expression = Expression('vm.int_value')
 
@@ -75,14 +75,14 @@ class InlineBindingTests:
 
 def test_bind_inline():
     bind = Mock()
-    context = BindingContext(
-        {
-            'node': Mock(node_globals = InheritedDict({'bind': bind, 'value': 1})),
-            'expression_body': 'bind()}:{value',
-            'setter': Mock(),
-            'xml_attr': XmlAttr('name')
-        }
-    )
+    context = BindingContext({
+        'node': Mock(node_globals = NodeGlobals({
+            'bind': bind, 'value': 1
+        })),
+        'expression_body': 'bind()}:{value',
+        'setter': Mock(),
+        'xml_attr': XmlAttr('name')
+    })
 
     actual = bind_inline(context)